requirements.txt
asgiref==3.2.10
asn1crypto==1.3.0
autopep8==1.5.4
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-extensions==3.0.4
django-nested-admin==3.3.2
django-print-sql==2018.3.6
django-rest-framework==0.1.0
django-xhtml2pdf==0.0.4
djangorestframework==3.11.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
mysqlclient==2.0.1
Pillow==7.1.2
pubnub==4.1.6
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.8
PyPDF2==1.26.0
Pyphen==0.9.5
python-decouple==3.3
python-monkey-business==1.0.0
pytz==2020.1
reportlab==3.5.47
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
tinycss2==1.0.2
toml==0.10.1
Unipath==1.1
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
